LayoutTests/fast/regions/get-regions-by-content-node-expected.txt

Test for 77746: [CSSRegions]Implement NamedFlow::getRegionsByContentNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regionList.length is 1
PASS regionList.item(0).id is "region"
PASS regionList.length is 2
PASS regionList.item(0).id is "region21"
PASS regionList.item(1).id is "region22"
PASS regionList.length is 0
PASS successfullyParsed is true

TEST COMPLETE
